refactor: move retriever to separate module.
